#8517 ui part of org.jkiss.dbeaver.ext.athena moved in org.jkiss.dbeaver.ext.athena.ui
