+++ plugins/org.jkiss.dbeaver.ext.athena.ui/META-INF/MANIFEST.MF
@@ -1,0 +1,17 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: %Bundle-Name
+Bundle-SymbolicName: org.jkiss.dbeaver.ext.athena.ui;singleton:=true
+Bundle-Version: 1.0.1.qualifier
+Bundle-Release-Date: 20200420
+Bundle-Activator: org.jkiss.dbeaver.ext.athena.ui.AthenaActivator
+Require-Bundle: org.eclipse.ui,
+ org.jkiss.dbeaver.ext.athena,
+ org.jkiss.dbeaver.ui.editors.connection,
+ org.jkiss.dbeaver.model,
+ org.jkiss.dbeaver.ui
+Bundle-ActivationPolicy: lazy
+Bundle-RequiredExecutionEnvironment: JavaSE-1.8
+Bundle-Vendor: %Bundle-Vendor
+Bundle-Localization: OSGI-INF/l10n/bundle
+Automatic-Module-Name: org.jkiss.dbeaver.ext.athena.ui
