+++ plugins/org.jkiss.dbeaver.ext.athena.ui/OSGI-INF/l10n/bundle.properties
@@ -1,0 +1,4 @@
+Bundle-Vendor = DBeaver Corp
+Bundle-Name = DBeaver Athena UI
+
+athena.dialog.connection.header=AWS Athena Connection Settings
