+++ plugins/org.jkiss.dbeaver.ext.athena.ui/OSGI-INF/l10n/bundle_de.properties
@@ -1,0 +1,1 @@
+athena.dialog.connection.header=AWS Athena Verbindungseinstellungen
