+++ plugins/org.jkiss.dbeaver.ext.athena.ui/OSGI-INF/l10n/bundle_pt_BR.properties
@@ -1,0 +1,4 @@
+Bundle-Name = DBeaver Athena UI
+Bundle-Vendor = DBeaver Corp
+
+athena.dialog.connection.header = Configura\u00E7\u00F5es de conex\u00E3o da AWS Athena
