+++ plugins/org.jkiss.dbeaver.ext.athena.ui/OSGI-INF/l10n/bundle_ru.properties
@@ -1,0 +1,1 @@
+athena.dialog.connection.header=AWS Athena \u043F\u0430\u0440\u0430\u043C\u0435\u0442\u0440\u044B \u043F\u043E\u0434\u043A\u043B\u044E\u0447\u0435\u043D\u0438\u044F
