+++ plugins/org.jkiss.dbeaver.ext.athena.ui/build.properties
@@ -1,0 +1,8 @@
+source.. = src/
+output.. = target/classes/
+bin.includes = plugin.xml,\
+               META-INF/,\
+               .,\
+               LICENSE.txt,\
+               OSGI-INF/,\
+               icons/,\
