+++ plugins/org.jkiss.dbeaver.ext.athena.ui/src/org/jkiss/dbeaver/ext/athena/ui/internal/AthenaMessages.properties
@@ -1,0 +1,9 @@
+label_region = Region
+label_connection = Connection
+label_access_key=AWS account access key ID
+label_access_key_id=AWS account secret access key ID
+label_aws_access_key = AWS Access Key
+label_s3_location=S3 location
+label_s3_output_location=S3 output location
+label_secret_key=AWS Secret Key
+label_security=Security
