+++ plugins/org.jkiss.dbeaver.ext.athena.ui/src/org/jkiss/dbeaver/ext/athena/ui/internal/AthenaMessages_de.properties
@@ -1,0 +1,9 @@
+label_region = Region
+label_connection = Verbindung
+label_access_key=AWS-Kontozugriffsschl\u00fcssel-ID
+label_access_key_id=AWS geheimer Zugangsschl\u00fcssel ID
+label_aws_access_key = AWS Zugangsschl\u00fcssel
+label_s3_location=S3 Standort
+label_s3_output_location=S3 Ausgangsort
+label_secret_key=AWS geheimer Schl\u00fcssel
+label_security=Sicherheit
