+++ plugins/org.jkiss.dbeaver.ext.athena.ui/src/org/jkiss/dbeaver/ext/athena/ui/internal/AthenaMessages_ru.properties
@@ -1,0 +1,9 @@
+label_region = \u0420\u0435\u0433\u0438\u043E\u043D
+label_connection = \u041F\u043E\u0434\u043A\u043B\u044E\u0447\u0435\u043D\u0438\u0435
+label_access_key=AWS \u0430\u043A\u043A\u0430\u0443\u043D\u0442 \u043A\u043B\u044E\u0447 \u0434\u043E\u0441\u0442\u0443\u043F\u0430 ID
+label_access_key_id=AWS \u0430\u043A\u043A\u0430\u0443\u043D\u0442 \u0441\u0435\u043A\u0440\u043A\u0442\u043D\u044B\u0439 \u043A\u043B\u044E\u0447 \u0434\u043E\u0441\u0442\u0443\u043F\u0430 ID
+label_aws_access_key = AWS \u043A\u043B\u044E\u0447 \u0434\u043E\u0441\u0442\u0443\u043F\u0430
+label_s3_location=S3 \u043C\u0435\u0441\u0442\u043E\u043F\u043E\u043B\u043E\u0436\u0435\u043D\u0438\u0435
+label_s3_output_location=S3 \u0440\u0430\u0441\u043F\u043E\u043B\u043E\u0436\u0435\u043D\u0438\u0435 \u0432\u044B\u0445\u043E\u0434\u043D\u044B\u0445 \u0434\u0430\u043D\u043D\u044B\u0445 
+label_security=\u0411\u0435\u0437\u043E\u043F\u0430\u0441\u043D\u043E\u0441\u0442\u044C
+label_secret_key=AWS \u0441\u0435\u043A\u0440\u0435\u0442\u043D\u044B\u0439 \u043A\u043B\u044E\u0447
