+++ plugins/org.jkiss.dbeaver.ext.athena/META-INF/MANIFEST.MF
@@ -4,14 +4,10 @@ Bundle-Name: DBeaver Athena extension
 Bundle-SymbolicName: org.jkiss.dbeaver.ext.athena;singleton:=true
 Bundle-Version: 1.0.37.qualifier
 Bundle-Release-Date: 20200420
-Require-Bundle: org.eclipse.ui,
- org.eclipse.core.runtime,
- org.jkiss.dbeaver.core,
- org.jkiss.dbeaver.model,
- org.jkiss.dbeaver.ext.generic
-Bundle-Activator: org.jkiss.dbeaver.ext.athena.AthenaActivator
+Require-Bundle: org.jkiss.dbeaver.ext.generic
 Bundle-ActivationPolicy: lazy
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Bundle-Vendor: DBeaver Corp
 Bundle-Localization: OSGI-INF/l10n/bundle
 Automatic-Module-Name: org.jkiss.dbeaver.ext.athena
+Export-Package: org.jkiss.dbeaver.ext.athena.model
