+++ plugins/org.jkiss.dbeaver.ext.athena/plugin.xml
@@ -50,23 +50,4 @@
         </dialect>
     </extension>
 
-    <extension point="org.jkiss.dbeaver.dataSourceView">
-        <view
-                id="org.jkiss.dbeaver.ext.snowflake.dataSourceWizardView"
-                targetID="org.jkiss.dbeaver.ext.ui.newConnectionWizard"
-                dataSource="athena"
-                label="%athena.dialog.connection.header"
-                class="org.jkiss.dbeaver.ext.athena.views.AthenaConnectionPage"
-                icon="icons/aws_athena_icon.png">
-        </view>
-        <view
-                class="org.jkiss.dbeaver.ext.athena.views.AthenaConnectionPage"
-                icon="icons/aws_athena_icon.png"
-                id="org.jkiss.dbeaver.ext.snowflake.dataSourceEditorView"
-                dataSource="athena"
-                label="%athena.dialog.connection.header"
-                targetID="org.jkiss.dbeaver.ext.ui.editConnectionDialog">
-        </view>
-    </extension>
-
 </plugin>
