+++ plugins/pom.xml
@@ -21,6 +21,7 @@
 
         <module>org.jkiss.dbeaver.ext.generic</module>
 
+        <module>org.jkiss.dbeaver.ext.athena</module>
         <module>org.jkiss.dbeaver.ext.clickhouse</module>
         <module>org.jkiss.dbeaver.ext.db2.zos</module>
         <module>org.jkiss.dbeaver.ext.db2.i</module>
@@ -70,7 +71,7 @@
                 <module>org.jkiss.dbeaver.ui.app.standalone</module>
                 <module>org.jkiss.dbeaver.ui.app.eclipse</module>
 
-                <module>org.jkiss.dbeaver.ext.athena</module>
+                <module>org.jkiss.dbeaver.ext.athena.ui</module>
                 <module>org.jkiss.dbeaver.ext.bigquery</module>
                 <module>org.jkiss.dbeaver.ext.exasol</module>
                 <module>org.jkiss.dbeaver.ext.db2</module>
